refactor: クラス化し別ファイルに切り出した
