+++ modules/data/src/main/java/net/pantasystem/milktea/data/di/module/UserModule.kt
@@ -93,4 +93,8 @@ abstract class UserModule {
     @Binds
     @Singleton
     internal abstract fun bindFollowApiAdapter(impl: FollowApiAdapterImpl): FollowApiAdapter
+
+    @Binds
+    @Singleton
+    internal abstract fun bindUserSuggestionApiAdapter(impl: UserSuggestionsApiAdapterImpl): UserSuggestionsApiAdapter
 }
