+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/user/UserRepositoryImpl.kt
@@ -3,39 +3,21 @@ package net.pantasystem.milktea.data.infrastructure.user
 
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
-import net.pantasystem.milktea.api.misskey.users.RequestUser
-import net.pantasystem.milktea.api.misskey.users.from
 import net.pantasystem.milktea.common.Logger
 import net.pantasystem.milktea.common.runCancellableCatching
-import net.pantasystem.milktea.common.throwIfHasError
 import net.pantasystem.milktea.common_android.hilt.IODispatcher
-import net.pantasystem.milktea.data.api.mastodon.MastodonAPIProvider
-import net.pantasystem.milktea.data.api.misskey.MisskeyAPIProvider
-import net.pantasystem.milktea.data.converters.UserDTOEntityConverter
-import net.pantasystem.milktea.data.infrastructure.notes.reaction.impl.history.ReactionHistoryDao
-import net.pantasystem.milktea.data.infrastructure.toUserRelated
-import net.pantasystem.milktea.model.account.AccountRepository
-import net.pantasystem.milktea.model.drive.FilePropertyDataSource
 import net.pantasystem.milktea.model.user.User
 import net.pantasystem.milktea.model.user.UserDataSource
 import net.pantasystem.milktea.model.user.UserNotFoundException
 import net.pantasystem.milktea.model.user.UserRepository
-import net.pantasystem.milktea.model.user.query.FindUsersFromFrequentlyReactionUsers
 import net.pantasystem.milktea.model.user.query.FindUsersQuery
-import net.pantasystem.milktea.model.user.query.FindUsersQuery4Mastodon
-import net.pantasystem.milktea.model.user.query.FindUsersQuery4Misskey
 import javax.inject.Inject
 
 internal class UserRepositoryImpl @Inject constructor(
-    val userDataSource: UserDataSource,
-    val filePropertyDataSource: FilePropertyDataSource,
-    val accountRepository: AccountRepository,
-    val misskeyAPIProvider: MisskeyAPIProvider,
-    val loggerFactory: Logger.Factory,
-    val userApiAdapter: UserApiAdapter,
-    private val mastodonAPIProvider: MastodonAPIProvider,
-    val userDTOEntityConverter: UserDTOEntityConverter,
-    private val reactionHistoryDao: ReactionHistoryDao,
+    private val userDataSource: UserDataSource,
+    private val loggerFactory: Logger.Factory,
+    private val userApiAdapter: UserApiAdapter,
+    private val userSuggestionsApiAdapter: UserSuggestionsApiAdapter,
     @IODispatcher val ioDispatcher: CoroutineDispatcher,
 ) : UserRepository {
     private val logger: Logger by lazy {
@@ -59,7 +41,7 @@ internal class UserRepositoryImpl @Inject constructor(
 
             if (localResult.getOrNull() == null) {
                 val user = userApiAdapter.show(userId, detail)
-                val result = userDataSource.add(user)
+                val result = userDataSource.add(user).getOrThrow()
                 logger.debug("add result: $result")
                 return@withContext userDataSource.get(userId).getOrThrow()
             }
@@ -99,24 +81,8 @@ internal class UserRepositoryImpl @Inject constructor(
         host: String?
     ) = runCancellableCatching<Unit> {
         withContext(ioDispatcher) {
-            val ac = accountRepository.get(accountId).getOrThrow()
-
-            when(val result = userApiAdapter.search(accountId, userName, host)) {
-                is SearchResult.Mastodon -> {
-                    userDataSource.addAll(
-                        result.users.map {
-                            it.toModel(ac)
-                        }
-                    )
-                }
-                is SearchResult.Misskey -> {
-                    result.users.forEach {
-                        userDTOEntityConverter.convert(ac, it, true).also { u ->
-                            userDataSource.add(u)
-                        }
-                    }
-                }
-            }
+            val users = userApiAdapter.search(accountId, userName, host)
+            userDataSource.addAll(users).getOrThrow()
         }
     }
 
@@ -134,51 +100,9 @@ internal class UserRepositoryImpl @Inject constructor(
 
     override suspend fun findUsers(accountId: Long, query: FindUsersQuery): List<User> {
         return withContext(ioDispatcher) {
-            val account = accountRepository.get(accountId).getOrThrow()
-            when(query) {
-                is FindUsersQuery4Mastodon.SuggestUsers -> {
-                    val api = mastodonAPIProvider.get(account)
-                    val body = requireNotNull(api.getSuggestionUsers(
-                        limit = query.limit
-                    ).throwIfHasError().body())
-                    val accounts = body.map {
-                        it.account
-                    }
-                    val relationships = requireNotNull(
-                        api.getAccountRelationships(ids = accounts.map { it.id })
-                            .throwIfHasError()
-                            .body()
-                    ).let { list ->
-                        list.associateBy {
-                            it.id
-                        }
-                    }
-                    val models = accounts.map {
-                        it.toModel(account, relationships[it.id]?.toUserRelated())
-                    }
-                    userDataSource.addAll(models).getOrThrow()
-                    models
-                }
-                is FindUsersQuery4Misskey -> {
-                    val request = RequestUser.from(query, account.token)
-                    val res = misskeyAPIProvider.get(account).getUsers(request)
-                        .throwIfHasError()
-                    res.body()?.map {
-                        userDTOEntityConverter.convert(account, it, true)
-                    }?.onEach {
-                        userDataSource.add(it)
-                    } ?: emptyList()
-                }
-                is FindUsersFromFrequentlyReactionUsers -> {
-                    val userIds = reactionHistoryDao.findFrequentlyReactionUserAndUnFollowed(
-                        accountId = accountId,
-                        limit = 20,
-                    ).map {
-                        it.targetUserId
-                    }
-                    userDataSource.getIn(accountId, userIds).getOrThrow()
-                }
-            }
+            val result = userSuggestionsApiAdapter.showSuggestions(accountId, query)
+            userDataSource.addAll(result).getOrThrow()
+            result
         }
     }
 
@@ -186,7 +110,7 @@ internal class UserRepositoryImpl @Inject constructor(
         return runCancellableCatching {
             withContext(ioDispatcher) {
                 val user = userApiAdapter.show(userId, true)
-                userDataSource.add(user)
+                userDataSource.add(user).getOrThrow()
             }
         }
     }
@@ -195,7 +119,7 @@ internal class UserRepositoryImpl @Inject constructor(
         return runCancellableCatching {
             withContext(ioDispatcher) {
                 val users = userApiAdapter.showUsers(userIds, true)
-                userDataSource.addAll(users)
+                userDataSource.addAll(users).getOrThrow()
                 users.map {
                     it.id
                 }
